Call add_special_courses whenever add_courses is invoked
The documentation does not explain that add_courses does not create
sufficient data for the app to run - in fact the Docker entrypoints (see
docker/web/start.sh) will populate the database with this data on every
run, with the result that the configuration steps are spread across the
docs, the app, and the build process.

Changing that is out of scope for now, but in the meantime have the data
setup script create the necessary courses.